浮点数可以正常工作了
使用浮点数进行方向控制中的运算，左右转向正常，但转向增益似乎还是大，下一步打算对偏差进行非线性的处理了

+++ src/app/calculation.h
@@ -9,7 +9,7 @@ extern  s16 Speed;
 extern	s16 SpeedSp;
 extern	s16 SpeedKc;
 extern  s16 DirectionKc;
-extern s16 DirectionErrorMan;
+extern float DirectionErrorMan;
 extern u8 ImageProFlag;
 
 void CCD_INC (void);
@@ -17,7 +17,7 @@ void DCMoto_init(void);
 void TURNPWM_init(void);
 void MaPan_init(void);
 s16 SpeedTransmitter(void);
-u32 DirectionPIDcontrol(s16 DirectionError);
+u32 DirectionPIDcontrol(float DirectionError);
 u32 SpeedPIcontrol(s16 Speed);
 void Control(void);
